Move vmitrait package to pkg/apimachinery
Relocate pkg/vmitrait to pkg/apimachinery/vmitrait and update
all import paths, BUILD.bazel references, and lint paths across
the codebase.

This is part of the effort to consolidate API machinery code
under pkg/apimachinery in preparation for extracting it into
an external module.
