+++ pkg/apimachinery/vmitrait/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["vmitrait.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/vmitrait",
+    importpath = "kubevirt.io/kubevirt/pkg/apimachinery/vmitrait",
     visibility = ["//visibility:public"],
     deps = ["//staging/src/kubevirt.io/api/core/v1:go_default_library"],
 )
