+++ pkg/hypervisor/kvm/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/hypervisor/kvm",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/apimachinery/vmitrait:go_default_library",
         "//pkg/downwardmetrics:go_default_library",
         "//pkg/hypervisor/common:go_default_library",
         "//pkg/tpm:go_default_library",
@@ -18,7 +19,6 @@ go_library(
         "//pkg/virt-handler/cgroup:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/mitchellh/go-ps:go_default_library",
