+++ pkg/virt-controller/services/BUILD.bazel
@@ -16,6 +16,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apimachinery:go_default_library",
+        "//pkg/apimachinery/vmitrait:go_default_library",
         "//pkg/config:go_default_library",
         "//pkg/container-disk:go_default_library",
         "//pkg/dra:go_default_library",
@@ -42,7 +43,6 @@ go_library(
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//pkg/virt-operator/util:go_default_library",
         "//pkg/virtiofs:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/export/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
