+++ pkg/virt-handler/launcher-clients/BUILD.bazel
@@ -9,11 +9,11 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-handler/launcher-clients",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/apimachinery/vmitrait:go_default_library",
         "//pkg/virt-handler/cache:go_default_library",
         "//pkg/virt-handler/cmd-client:go_default_library",
         "//pkg/virt-handler/isolation:go_default_library",
         "//pkg/virt-handler/notify-server/pipe:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/golang.org/x/sync/singleflight:go_default_library",
