+++ pkg/virt-handler/launcher-clients/launcher-clients.go
@@ -30,11 +30,11 @@ import (
 	v1 "kubevirt.io/api/core/v1"
 	"kubevirt.io/client-go/log"
 
+	"kubevirt.io/kubevirt/pkg/apimachinery/vmitrait"
 	virtcache "kubevirt.io/kubevirt/pkg/virt-handler/cache"
 	cmdclient "kubevirt.io/kubevirt/pkg/virt-handler/cmd-client"
 	"kubevirt.io/kubevirt/pkg/virt-handler/isolation"
 	"kubevirt.io/kubevirt/pkg/virt-handler/notify-server/pipe"
-	"kubevirt.io/kubevirt/pkg/vmitrait"
 )
 
 type LauncherClientsManager interface {
