+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/apimachinery/vmitrait:go_default_library",
         "//pkg/apimachinery/wait:go_default_library",
         "//pkg/cloud-init:go_default_library",
         "//pkg/config:go_default_library",
@@ -76,7 +77,6 @@ go_library(
         "//pkg/virt-launcher/virtwrap/statsconv:go_default_library",
         "//pkg/virt-launcher/virtwrap/storage:go_default_library",
         "//pkg/virt-launcher/virtwrap/util:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/backup/v1alpha1:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/plugin/v1alpha1:go_default_library",
