+++ tests/libdomain/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "kubevirt.io/kubevirt/tests/libdomain",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/apimachinery/vmitrait:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//tests/exec:go_default_library",
